requirements.txt
python-chess>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
flask>=3.1.0
